Fix use TLS protocol in .NET461
