+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Ingestion/Http/HttpNetworkAdapter.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Security.Authentication;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -16,7 +15,7 @@ namespace Microsoft.AppCenter.Ingestion.Http
     {
         internal const string ContentTypeValue = "application/json; charset=utf-8";
 
-        // used by the windows platform to override and force tls1.2
+        // Prepare an HTTP message handler for configuring the TLS protocol.
         internal static readonly Func<HttpMessageHandler> HttpMessageHandlerOverride;
 
         private HttpClient _httpClient;
