+++ SDK/AppCenter/Microsoft.AppCenter.WindowsDesktop/Ingestion/Http/HttpNetworkAdapter.cs
@@ -2,7 +2,12 @@
 // Licensed under the MIT License.
 
 using System.Net.Http;
+
+#if NET461
+using System.Net;
+#else
 using System.Security.Authentication;
+#endif
 
 namespace Microsoft.AppCenter.Ingestion.Http
 {
@@ -12,10 +17,19 @@ namespace Microsoft.AppCenter.Ingestion.Http
         // Static initializer specific to windows desktop platforms.
         static HttpNetworkAdapter()
         {
+#if NET461
+            // ReSharper disable once InvertIf
+            if ((ServicePointManager.SecurityProtocol & SecurityProtocolType.Tls12) != SecurityProtocolType.Tls12)
+            {
+                ServicePointManager.SecurityProtocol |= SecurityProtocolType.Tls12;
+                AppCenterLog.Debug(AppCenterLog.LogTag, "Enabled TLS 1.2 explicitly as it was disabled.");
+            }
+#else
             HttpMessageHandlerOverride = () => new HttpClientHandler
             {
                 SslProtocols = SslProtocols.Tls12
             };
+#endif
         }
     }
 }
