+++ Tests/Microsoft.AppCenter.Test.Windows/Ingestion/Http/HttpNetworkAdapterTest.cs
@@ -3,9 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Net;
 using System.Net.Http;
-using System.Security.Authentication;
 using Microsoft.AppCenter.Ingestion.Http;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -37,21 +35,6 @@ namespace Microsoft.AppCenter.Test.Windows.Ingestion.Http
             Assert.IsTrue(request.Headers.Contains(IngestionHttp.AppSecret));
             Assert.IsTrue(request.Headers.Contains(IngestionHttp.InstallId));
             Assert.IsInstanceOfType(request.Content, typeof(StringContent));
-        }
-
-        [TestMethod]
-        public void EnableTls12WhenServicePointManagerSetToSystemDefault()
-        {
-            ServicePointManager.SecurityProtocol = SecurityProtocolType.SystemDefault;
-
-            var httpClientHandler = HttpNetworkAdapter.HttpMessageHandlerOverride() as HttpClientHandler;
-
-            Assert.IsNotNull(httpClientHandler);
-
-            Assert.AreEqual(SslProtocols.Tls12, httpClientHandler.SslProtocols);
-
-            // Just check no side effect.
-            Assert.AreEqual(SecurityProtocolType.SystemDefault, ServicePointManager.SecurityProtocol);
         }
     }
 }
